refactor: 패키지 이동 및 커스텀 예외 추가
