+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -6,7 +6,6 @@ import org.springframework.http.ProblemDetail;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import roomescape.global.ReservationException;
 
 @Slf4j
 @ControllerAdvice
@@ -22,5 +21,11 @@ public class GlobalExceptionHandler {
     public ProblemDetail handleReservationException(ReservationException e) {
         log.error("예외 발생: {}", e.getStackTrace(), e);
         return ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, e.getMessage());
+    }
+
+    @ExceptionHandler(UnAuthorizedException.class)
+    public ProblemDetail handleUnAuthorizedException(UnAuthorizedException e) {
+        log.error("예외 발생: {}", e.getStackTrace(), e);
+        return ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, "인증 정보를 찾을 수 없습니다.");
     }
 }
