+++ src/main/java/roomescape/exception/UnAuthorizedException.java
@@ -1,0 +1,10 @@
+package roomescape.exception;
+
+public class UnAuthorizedException extends RuntimeException {
+    public UnAuthorizedException(String message) {
+        super(message);
+    }
+
+    public UnAuthorizedException() {
+    }
+}
